Fixed images

+++ app/src/main/java/pt/unl/fct/mealroullete/homepage/history/DailyHistoryFragment.kt
@@ -34,7 +34,7 @@ class DailyHistoryFragment : Fragment() {
             totalCalories += ingredient.calories
         }
 
-        child.findViewById<ImageView>(R.id.historyImage).setImageResource(recipe.image)
+        child.findViewById<ImageView>(R.id.historyImage).setImageResource((recipe.image as Int))
         child.findViewById<TextView>(R.id.historyName).text = recipe.name
         child.findViewById<TextView>(R.id.historyCalories).text = recipe.calories.toString()
         child.setOnClickListener {
